Apple metal rendering loop improvements
In the past, NanoGUI acquired a drawable at the beginning of each frame,
rendered to the underlying texture object, and finally called
``[drawable present]``. This was a bad design: ``[drawable present]``
call does not respect the ordering of the previously scheduled work,
which can lead to garbage being displayed (which DID happen sometimes).
The proper way to do this is to submit the ``present`` operation using a
command buffer. The implementation also needs to wait for the command
buffer to be flushed before shutting down the app, or error messages
will be generated sometimes (this is fixed as well).

I also experimented with further variants of the loop that render to an
texture, acquire the drawable briefly at the end to blit the texture.
This was based on a recommendation by Apple that it is preferable to
hold the drawable only for a very short amount of time. However, this
more complex approach made no difference on performance/latency, likely
because the NanoGUI applications are all extremely lightweight. I left
the current approach (potentially to be revisited in the future).

One thing I find quite noticeable is that the mouse-to-photon lag is
*much* lower in examples that use ``Lazy`` rendering mode, e.g., when
moving a window. VSync is turned off in that mode, and the application
just deals with events as they come.

This turns out to be the case because some Macs have ProMotion (120hz)
displays. Not waiting for VSync makes it possible to benefit from this
faster refresh rate. Unfortunately there doesn't seem to be a
straightforward way to opt into 120Hz *while* having VSync enabled. It
seems that doing so requires a very different set of Cocoa APIs
(CVDisplayLink or the newer CAMetalDisplayLink) that take control of the
main loop and look rather incompatible with how we currently use GLFW
for this.

Thus RunMode::Lazy is a good choice if supported by the application.