Cache the preferred size of widgets
NanoGUI potentially spends a significant amount of runtime in the
functions ``nvgTextBounds()`` and ``nvgTextBoxBounds()`` that are used
to compute the preferred size of widgets.

This commit adopts a strategy that caches the preferred size in
bottom-level widgets (i.e., ones that don't layout a set of child
widgets).

The main complication is that various kinds of changes (setting
different captions, themes, etc.) can invalidate the precomputed
preferred size, so a relatively large number of setters need touchups.