Fix tooltip rendering
When rendering in lazy mode, tooltips fail to appear at the moment due
to a regression. This would require a notification 200msec after the
last mouse motion so that the screen can be re-rendered and show the
tooltip.

This commit adds a timer thread that serves this purpose. The
implementation uses condition variables to sleep until a deadline, which
can be further extended as events arrive. This is designed to minimize
wakeups and inter-thread communication.